MAINT: moved annotation tests in test_seq_annotation.py
