+++ tests/test_core/test_sequence.py
@@ -10,7 +10,6 @@ import pytest
 import cogent3
 from cogent3._version import __version__
 from cogent3.core import alphabet as c3_alphabet
-from cogent3.core import annotation_db as anndb_module
 from cogent3.core import genetic_code as c3_genetic_code
 from cogent3.core import moltype as c3_moltype
 from cogent3.core import sequence as c3_sequence
@@ -892,74 +891,6 @@ def test_strand_symmetry():
     assert len(ssym.observed.keys()) <= 8
     assert numpy.allclose(ssym.observed["AA"].to_array(), [2, 1])
     assert numpy.allclose(ssym.observed["CC"].to_array(), [1, 2])
-
-
-def test_is_annotated():
-    """is_annotated operates correctly"""
-    s = c3_moltype.DNA.make_seq(seq="ACGGCTGAAGCGCTCCGGGTTTAAAACG", name="s1")
-    _ = s.add_feature(biotype="gene", name="blah", spans=[(0, 10)])
-    assert s.is_annotated()
-
-
-@pytest.mark.parametrize("biotype", ["gene", "exon", ("gene", "exon")])
-def test_is_annotated_biotype(biotype):
-    """is_annotated operates correctly"""
-    s = c3_moltype.DNA.make_seq(seq="ACGGCTGAAGCGCTCCGGGTTTAAAACG", name="s1")
-    _ = s.add_feature(biotype="gene", name="blah", spans=[(0, 10)])
-    _ = s.add_feature(biotype="exon", name="blah", spans=[(0, 10)])
-    assert s.is_annotated(biotype=biotype)
-
-
-def test_annotation_db_lazy_evaluation():
-    s = c3_moltype.DNA.make_seq(seq="AC", name="s1")
-    assert isinstance(s._annotation_db, list)
-    # now if we invoke the property we get an actual db instance created
-    assert isinstance(s.annotation_db, anndb_module.SupportsFeatures)
-
-
-def test_init_with_annotationdb():
-    anndb = anndb_module.GffAnnotationDb()
-    s = c3_moltype.DNA.make_seq(seq="AC", name="s1", annotation_db=anndb)
-    assert isinstance(s.annotation_db, anndb_module.GffAnnotationDb)
-    assert s.annotation_db is anndb
-
-
-def test_init_with_annotation_offset():
-    s = c3_moltype.DNA.make_seq(seq="AC", name="s1", annotation_offset=2)
-    assert s.annotation_offset == 2
-
-
-def test_init_with_annotation_offset_sliced():
-    s = c3_moltype.DNA.make_seq(seq="ACTTTGG", name="s1", annotation_offset=2)
-    sl = s[2:5]
-    assert sl.annotation_offset == 4
-    _, start, stop, _ = sl.parent_coordinates()
-    assert start == 4
-    assert stop == 7
-
-
-def test_not_is_annotated():
-    """is_annotated operates correctly"""
-    s = c3_moltype.DNA.make_seq(seq="ACGGCTGAAGCGCTCCGGGTTTAAAACG", name="s1")
-    assert not s.is_annotated()
-    # annotation on different seq
-    s.annotation_db.add_feature(
-        seqid="s2",
-        biotype="gene",
-        name="blah",
-        spans=[(0, 10)],
-    )
-    assert not s.is_annotated()
-    # annotation wrong biotype
-    s.annotation_db.add_feature(
-        seqid="s1",
-        biotype="exon",
-        name="blah",
-        spans=[(0, 10)],
-    )
-    assert not s.is_annotated(biotype="gene")
-    s.annotation_db = None
-    assert not s.is_annotated()
 
 
 def test_to_html():
